requirements.txt
stegoveritas_binwalk==2.1.2
binwalk==2.1.0
pycryptodome==3.9.9
